Refactor memory management system by introducing vector database manager
This commit refactors the memory management system by replacing the existing DBManager with a new _VectorDBManager class, which enhances the handling of vector embeddings. The embedding functions have been restructured to utilize the SentenceTransformer model for generating embeddings, and the embedding configuration has been updated accordingly. Additionally, the embedding.py file has been removed as its functionality has been integrated into api.py. A new performance test for package import time has also been added to ensure efficiency.

Changes include:
- Introduced _VectorDBManager for improved vector database management.
- Integrated SentenceTransformer for embedding generation.
- Removed embedding.py as its functionality is now in api.py.
- Added performance test for import time.

This refactor aims to streamline the memory management process and improve overall performance.
